+++ SharePointFramework/PortfolioWebParts/src/webparts/portfolioOverview/index.ts
@@ -18,9 +18,12 @@ import {
   PropertyFieldCollectionData
 } from '@pnp/spfx-property-controls/lib/PropertyFieldCollectionData'
 import _ from 'lodash'
-import { PortalDataServiceDefaultConfiguration, UserMessage, ErrorWithIntent } from 'pp365-shared-library'
+import {
+  PortalDataServiceDefaultConfiguration,
+  UserMessage,
+  ErrorWithIntent
+} from 'pp365-shared-library'
 import { DataAdapter } from 'data'
-
 
 export default class PortfolioOverviewWebPart extends BasePortfolioWebPart<IPortfolioOverviewProps> {
   private _configuration: IPortfolioOverviewConfiguration
@@ -29,29 +32,41 @@ export default class PortfolioOverviewWebPart extends BasePortfolioWebPart<IPort
 
   public render(): void {
     if (!this._configuration) {
-      render(createElement(UserMessage, {
-        title: this._error?.name,
-        text: this._error?.message,
-        intent: this._error?.intent
-      }), this.domElement)
+      render(
+        createElement(UserMessage, {
+          title: this._error?.name,
+          text: this._error?.message,
+          intent: this._error?.intent
+        }),
+        this.domElement
+      )
       return
     }
     this.renderComponent<IPortfolioOverviewProps>(PortfolioOverview, {
       configuration: this._configuration,
       onSetPortfolio: this.setPortfolio.bind(this),
-      selectedPortfolioId: this._selectedPortfolioId
+      selectedPortfolioId: this._selectedPortfolioId,
+      selectedPortfolio: this.properties.portfolios.find(
+        ({ uniqueId }) => uniqueId === this._selectedPortfolioId
+      )
     })
   }
 
   /**
    * Callback function to set the portfolio to display in the web part.
-   * 
+   *
    * @param portfolioId Unique ID of the portfolio
    */
   private async setPortfolio(portfolioId: string): Promise<void> {
     this._selectedPortfolioId = portfolioId
-    const portfolio = this.properties.portfolios.find(({ uniqueId }) => uniqueId === this._selectedPortfolioId)
-    this.dataAdapter = await new DataAdapter(this.context, this.sp).configure(this.context, null, portfolio)
+    const portfolio = this.properties.portfolios.find(
+      ({ uniqueId }) => uniqueId === this._selectedPortfolioId
+    )
+    this.dataAdapter = await new DataAdapter(this.context, this.sp).configure(
+      this.context,
+      null,
+      portfolio
+    )
     this._configuration = await this.dataAdapter.getPortfolioConfig()
     this.render()
   }
@@ -59,7 +74,9 @@ export default class PortfolioOverviewWebPart extends BasePortfolioWebPart<IPort
   public async onInit(): Promise<void> {
     try {
       this._selectedPortfolioId = this.properties.selectedPortfolioId
-      const portfolio = this.properties.portfolios.find(({ uniqueId }) => uniqueId === this._selectedPortfolioId)
+      const portfolio = this.properties.portfolios.find(
+        ({ uniqueId }) => uniqueId === this._selectedPortfolioId
+      )
       await super.onInit(portfolio)
       this._configuration = await this.dataAdapter.getPortfolioConfig()
     } catch (error) {
@@ -73,7 +90,9 @@ export default class PortfolioOverviewWebPart extends BasePortfolioWebPart<IPort
    *
    * @param targetProperty Target property
    */
-  protected _getOptions(targetProperty: keyof IPortfolioOverviewProps): IPropertyPaneDropdownOption[] {
+  protected _getOptions(
+    targetProperty: keyof IPortfolioOverviewProps
+  ): IPropertyPaneDropdownOption[] {
     switch (targetProperty) {
       case 'portfolios': {
         return this.properties.portfolios.map((portfolio) => ({
@@ -106,12 +125,11 @@ export default class PortfolioOverviewWebPart extends BasePortfolioWebPart<IPort
                 PropertyPaneToggle('showSearchBox', {
                   label: strings.ShowSearchBoxLabel
                 }),
-                !_.isEmpty(this.properties.portfolios) && (
+                !_.isEmpty(this.properties.portfolios) &&
                   PropertyPaneDropdown('selectedPortfolioId', {
                     label: strings.SelectedPortfolioLabel,
                     options: this._getOptions('portfolios')
-                  })
-                ),
+                  }),
                 PropertyPaneDropdown('defaultViewId', {
                   label: strings.DefaultViewLabel,
                   options: this._getOptions('defaultViewId')
@@ -131,23 +149,22 @@ export default class PortfolioOverviewWebPart extends BasePortfolioWebPart<IPort
                   label: strings.ShowExcelExportButtonLabel
                 }),
                 this.properties.showExcelExportButton &&
-                PropertyPaneToggle('includeViewNameInExcelExportFilename', {
-                  label: strings.IncludeViewNameInExcelExportFilenameLabel
-                }),
-                !_.isEmpty(this.properties.portfolios) && (
+                  PropertyPaneToggle('includeViewNameInExcelExportFilename', {
+                    label: strings.IncludeViewNameInExcelExportFilenameLabel
+                  }),
+                !_.isEmpty(this.properties.portfolios) &&
                   PropertyPaneToggle('showPortfolioSelector', {
                     label: strings.ShowPortfolioSelectorLabel,
                     onText: strings.ShowPortfolioSelectorOnText,
                     offText: strings.ShowPortfolioSelectorOffText
-                  })
-                ),
+                  }),
                 PropertyPaneToggle('showViewSelector', {
                   label: strings.ShowViewSelectorLabel
                 }),
                 this.properties.showViewSelector &&
-                PropertyPaneToggle('showProgramViews', {
-                  label: strings.ShowProgramViewsLabel
-                })
+                  PropertyPaneToggle('showProgramViews', {
+                    label: strings.ShowProgramViewsLabel
+                  })
               ].filter(Boolean)
             },
             {
@@ -164,34 +181,38 @@ export default class PortfolioOverviewWebPart extends BasePortfolioWebPart<IPort
                       id: 'title',
                       title: strings.TitleLabel,
                       type: CustomCollectionFieldType.string,
-                      required: true,
+                      required: true
                     },
                     {
                       id: 'url',
                       title: strings.UrlFieldLabel,
                       type: CustomCollectionFieldType.string,
-                      required: true,
+                      required: true
                     },
                     {
                       id: 'viewsListName',
                       title: strings.ViewsListNameFieldLabel,
                       type: CustomCollectionFieldType.string,
-                      defaultValue: PortalDataServiceDefaultConfiguration?.listNames?.PORTFOLIO_VIEWS,
-                      required: true,
+                      defaultValue:
+                        PortalDataServiceDefaultConfiguration?.listNames?.PORTFOLIO_VIEWS,
+                      required: true
                     },
                     {
                       id: 'columnsListName',
                       title: strings.ColumnsListNameFieldLabel,
                       type: CustomCollectionFieldType.string,
-                      defaultValue: PortalDataServiceDefaultConfiguration?.listNames?.PROJECT_COLUMNS,
-                      required: true,
+                      defaultValue:
+                        PortalDataServiceDefaultConfiguration?.listNames?.PROJECT_COLUMNS,
+                      required: true
                     },
                     {
                       id: 'columnConfigListName',
                       title: strings.ColumnConfigListNameFieldLabel,
                       type: CustomCollectionFieldType.string,
-                      defaultValue: PortalDataServiceDefaultConfiguration?.listNames?.PROJECT_COLUMN_CONFIGURATION,
-                      required: true,
+                      defaultValue:
+                        PortalDataServiceDefaultConfiguration?.listNames
+                          ?.PROJECT_COLUMN_CONFIGURATION,
+                      required: true
                     }
                   ]
                 })
